Add files via upload
The Lulu Hypermarket Voucher System is a  Java Program designed to simulate a process of customers entering their details for a promotional offers. The program allows the customers to fill out vouchers and randomly pick 6 Lucky Winners from the list of participants. The application demonstrates key programming concepts such as Object-Oriented Programming(OOP).